Fix #66 - Refactor the web-service implementation
