+++ src/server/controllers/indexController.js
@@ -1,0 +1,7 @@
+const VERSION = require('../../../package.json').version
+
+const getIndex = (ctx) => {
+  ctx.ok({module: ' fTbus', VERSION})
+}
+
+module.exports = { getIndex }
