+++ src/server/controllers/infoController.js
@@ -1,4 +1,5 @@
 const getInfo = (ctx) => {
   ctx.ok({ config: ctx.app.engine.config })
 }
+
 module.exports = { getInfo }
