+++ src/server/middlewares/auth.js
@@ -24,7 +24,7 @@ const auth = (opts = {}) => {
     if (user && user.pass && user.name === opts.name) {
       const hash = crypto.createHash('sha256').update(user.pass).digest('hex')
       if (hash === opts.pass) return next()
-      console.error('bad hash', hash)
+      ctx.app.engine.logger.error('bad hash', hash)
     }
     return ctx.throw(401, null, { headers: { 'WWW-Authenticate': `Basic realm="${opts.realm.replace(/"/g, '\\"')}"` } })
   }
